created persistent homepage logic
