+++ models/user.js
@@ -20,7 +20,9 @@ module.exports = function(sequelize, DataTypes) {
       defaultValue: 2.5
     },
     userName: {
-      type: DataTypes.STRING
+      type: DataTypes.STRING,
+      unique: true,
+      allowNull: false
     },
     password: {
       type: DataTypes.STRING,
