+++ routes/htmlRoutes.js
@@ -1,7 +1,13 @@
 var db = require("../models");
-var Handlebars = require("express-handlebars");
+
 // Requiring our custom middleware for checking if a user is logged in
 var isAuthenticated = require("../config/middleware/isAuthenticated");
+
+// Custom Middleware to create dynamic gradient
+require("../config/middleware/gradientSort");
+
+// Custom Middleware to sort Index by Vote Sum, to total Vote Counts, to implement frontend Vote displays
+require("../config/middleware/voteSortLogic");
 
 // ==== Debugging Functions ==== //
 // Log Sequelize Statements as raw SQL
@@ -23,77 +29,6 @@ logRankAndColor = function(results) {
   }
 };
 
-// ==== Handlebars Helpers ==== //
-plusMinusVoteCount = function(VotesObj) {
-  var negativeVoteCount = 0;
-  var positiveVoteCount = 0;
-  for (var i = 0; i < VotesObj.length; i++) {
-    var currentVoteVal = VotesObj[i].voteValue;
-
-    if (currentVoteVal > 0) {
-      positiveVoteCount++;
-    } else {
-      negativeVoteCount++;
-    }
-  }
-  return (
-    '<span id="up-vote__count">' +
-    positiveVoteCount +
-    '</span> / <span id="down-vote__count">' +
-    negativeVoteCount +
-    "</span>"
-  );
-};
-
-// Creat Color Constructor to add to returned Sequelize Model
-function Color(hslObject) {
-  this.hue = hslObject.hue;
-  this.saturation = hslObject.saturation;
-  this.lightness = hslObject.lightness;
-}
-
-// Sort Sequelize Query by the Items' voteScore
-sortByItemScoreSum = function(results) {
-  results.sort(function(a, b) {
-    return b.dataValues.itemScore - a.dataValues.itemScore;
-  });
-};
-
-// Calculate the rate of Color change from top color to bottom color
-calcColorChange = function(topColor, bottomColor, listLength) {
-  var colorDistance = {
-    hue: topColor.hue - bottomColor.hue,
-    saturation: topColor.saturation - bottomColor.saturation,
-    lightness: topColor.lightness - bottomColor.lightness
-  };
-  var colorChangePerItem = {
-    hue: Math.floor(colorDistance.hue / (listLength - 1)),
-    saturation: Math.floor(colorDistance.saturation / (listLength - 1)),
-    lightness: Math.floor(colorDistance.lightness / (listLength - 1))
-  };
-  return colorChangePerItem;
-};
-
-// Set the div's item color based on it position relative to the topColor/topItem
-setItemColor = function(startColor, hslValue, itemRank) {
-  return {
-    hue: startColor.hue - hslValue.hue * itemRank,
-    saturation: startColor.saturation - hslValue.saturation * itemRank,
-    lightness: startColor.lightness - hslValue.lightness * itemRank
-  };
-};
-
-insertItemColorVal = function(results) {
-  // Set topColor and bottomColor values for gradient change
-  var greenest = { hue: 130, saturation: 87, lightness: 45 };
-  var reddest = { hue: 10, saturation: 87, lightness: 45 };
-  hslDeltaVal = calcColorChange(greenest, reddest, results.length);
-  for (var i = 0; i < results.length; i++) {
-    itemColor = setItemColor(greenest, hslDeltaVal, i);
-    results[i].Color = new Color(itemColor);
-  }
-};
-
 module.exports = function(app) {
   app.get("/unauth", function(req, res) {
     db.Vote.findAll({
@@ -112,13 +47,15 @@ module.exports = function(app) {
       });
     });
   });
+
   // Load homepage
   app.get("/", isAuthenticated, function(req, res) {
     // object destructuring: this miracle syntax allows you to ref
     // "userName" as if you were referencing "req.user.userName";
     // saves keystrokes.
-    var { userName, bio, image, begScore, last_login, name } = req.user;
+    var { id, userName, bio, image, begScore, last_login, name } = req.user;
     var userData = {
+      id: id,
       userName: userName,
       bio: bio,
       image: image,
@@ -127,7 +64,7 @@ module.exports = function(app) {
       name: name
     };
 
-    console.log(userData);
+    // console.log(userData.id);
 
     db.Vote.findAll({
       attributes: [
@@ -140,51 +77,41 @@ module.exports = function(app) {
     }).then(function(dbItems) {
       sortByItemScoreSum(dbItems);
       insertItemColorVal(dbItems);
-      res.render("index", {
-        items: dbItems,
-        helpers: {
-          plusMinusVoteCount: plusMinusVoteCount
-        }
+      db.Vote.findAll({
+        where: { UserId: userData.id }
+      }).then(function(userVotes) {
+        res.render("index", {
+          items: dbItems,
+          userInfo: userVotes,
+          helpers: {
+            plusMinusVoteCount: plusMinusVoteCount,
+            applySelected: applySelected
+          }
+        });
       });
     });
   });
 
   //load items page (Beg input and suggested items list)
   app.get("/items", function(req, res) {
-    db.Item.findAll({}).then(function(dbItem) {
-      res.render("items");
-    });
+    res.render("items");
   });
 
   app.get("/test-modal", function(req, res) {
-    // db.User.findAll({}).then(function(dbItems) {
-    res.render("test", {
-      // msg: "Welcome!",
-      // examples: dbItems
-    });
-    // console.log(dbItems);
-    // });
+    res.render("test");
   });
 
   // Load Item Template - pass db Item via ItemId
   app.get("/items/:id", function(req, res) {
     db.Item.findAll({
-      where: {
-        id: req.params.id
-      },
-      include: [
-        {
-          model: db.User
-        },
-        {
-          model: db.Vote
-        }
-      ]
+      where: { id: req.params.id },
+      include: [{ model: db.User }, { model: db.Vote }]
     }).then(function(dbItem) {
       res.render("item-single", {
         item: dbItem,
         helpers: {
-          plusMinusVoteCount: plusMinusVoteCount
+          plusMinusVoteCount: plusMinusVoteCount,
+          applySelected: applySelected
         }
       });
     });
@@ -192,6 +119,17 @@ module.exports = function(app) {
 
   // Load User Template - pass db Item via ItemId
   app.get("/user/:id", function(req, res) {
+    var { id, userName, bio, image, begScore, last_login, name } = req.user;
+    var userData = {
+      id: id,
+      userName: userName,
+      bio: bio,
+      image: image,
+      begScore: begScore,
+      last_login: last_login,
+      name: name
+    };
+
     db.User.findAll({
       where: {
         id: req.params.id
@@ -201,7 +139,8 @@ module.exports = function(app) {
       res.render("user-single", {
         user: dbUser,
         helpers: {
-          plusMinusVoteCount: plusMinusVoteCount
+          plusMinusVoteCount: plusMinusVoteCount,
+          applySelected: applySelected
         }
       });
     });
